requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0

# OCR Engines - Multiple engines for ensemble approach
paddleocr>=2.7.0
pytesseract>=0.3.10
easyocr>=1.7.0

# PDF Processing
pymupdf>=1.27.0

# Image Processing & Computer Vision
numpy>=1.26.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
pyyaml>=6.0

# Async
aiofiles>=23.0
